Cleaning up script and supporting input type
